+++ src/components/DevTools.jsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const DevTools = () => <span />
+
+export default DevTools