+++ src/components/Search.jsx
@@ -2,16 +2,12 @@ import React from 'react'
 import { render } from 'react-dom'
 import PureRenderMixin from 'react-addons-pure-render-mixin'
 import MUI from 'material-ui'
-import SvgIcons from 'material-ui/lib/svg-icons'
+import ActionSearch from 'material-ui/lib/svg-icons/action/search'
 
 const {
   TextField,
   Paper,
 } = MUI
-
-const {
-  ActionSearch,
-} = SvgIcons
 
 const Search = props => (
   <Paper style={Object.assign({}, styles.search, props.containerStyle)} zDepth={1}>