+++ src/index.jsx
@@ -3,15 +3,29 @@ import './styles'
 import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
+import { Router, Route, Redirect, IndexRoute, hashHistory } from 'react-router'
 import injectTapEventPlugin from 'react-tap-event-plugin'
 import store from './store'
-import Root from './containers/Root'
+import DevTools from './components/DevTools'
+import App from './containers/App'
+import About from './containers/About'
+import Sites from './containers/Sites'
 
 injectTapEventPlugin()
 
 render(
   <Provider store={store}>
-    <Root />
+    <div>
+      <Router history={hashHistory}>
+        <Route path='/' component={App}>
+          <Redirect from='/' to='/sites' />
+          <Route path='sites' component={Sites} />
+          <Route path='about' component={About} />
+          <IndexRoute component={Sites}/>
+        </Route>
+      </Router>
+      <DevTools />
+    </div>
   </Provider>,
   document.getElementById('root')
 )